separate DSL declaration and interpreter from syntax
